+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log for the search node
 
+## Development
+
+* Support for geo-point in mapping configuration.
+* Added create new index API call.
+* Added remove index API call.
+* Change error message to be JSON.
+* Fixed error in addIndex where "Mappings already exists." is always returned.
+
 ## 2.1.3
 
 * Fixed sort error to now use multi field .sort in the index, which is not analyzed.